entity "Photo" was renamed to "CarPhoto"
