+++ CarCatalog.Domain/Entities/CarPhoto.cs
@@ -2,10 +2,9 @@ using CarCatalog.Domain.Common.BaseEntities;
 
 namespace CarCatalog.Domain.Entities;
 
-public class Photo : AuditableBaseEntity
+public class CarPhoto : AuditableBaseEntity
 {
     public Guid CarId { get; set; }
     public Car Car { get; set; }
-    
-    public string PhotoURL { get; set; }
+    public List<string> PhotoURL { get; set; } = new List<string>();
 }
